@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2020 Martin Donath <martin.donath@squidfunk.com>
+ * Copyright (c) 2016-2021 Martin Donath <martin.donath@squidfunk.com>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to
@@ -23,5 +23,5 @@
 export * from "./_"
 export * from "./focus"
 export * from "./offset"
-export * from "./select"
+export * from "./selection"
 export * from "./size"
