@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2020 Martin Donath <martin.donath@squidfunk.com>
+ * Copyright (c) 2016-2021 Martin Donath <martin.donath@squidfunk.com>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to
@@ -29,7 +29,6 @@ import {
   SearchHighlightFactoryFn,
   setupSearchHighlighter
 } from "../highlighter"
-import { SearchOptions } from "../options"
 import {
   SearchQueryTerms,
   getSearchQueryTerms,
@@ -57,19 +56,34 @@ export interface SearchIndexDocument {
   text: string                         /* Document text */
 }
 
+/* ------------------------------------------------------------------------- */
+
+/**
+ * Search index pipeline function
+ */
+export type SearchIndexPipelineFn =
+  | "trimmer"                          /* Trimmer */
+  | "stopWordFilter"                   /* Stop word filter */
+  | "stemmer"                          /* Stemmer */
+
+/**
+ * Search index pipeline
+ */
+export type SearchIndexPipeline = SearchIndexPipelineFn[]
+
+/* ------------------------------------------------------------------------- */
+
 /**
  * Search index
  *
  * This interfaces describes the format of the `search_index.json` file which
- * is automatically built and provided by the MkDocs search plugin. However,
- * note that Material for MkDocs adds some further configuration options for
- * search, which are added as `options` by the theme.
+ * is automatically built by the MkDocs search plugin.
  */
 export interface SearchIndex {
   config: SearchIndexConfig            /* Search index configuration */
   docs: SearchIndexDocument[]          /* Search index documents */
   index?: object                       /* Prebuilt index */
-  options: SearchOptions               /* Search options */
+  pipeline?: SearchIndexPipeline       /* Search index pipeline */
 }
 
 /* ------------------------------------------------------------------------- */
@@ -85,24 +99,9 @@ export interface SearchMetadata {
 /* ------------------------------------------------------------------------- */
 
 /**
- * Search result document
- */
-export type SearchResultDocument = SearchDocument & SearchMetadata
-
-/**
- * Search result item
- */
-export type SearchResultItem = SearchResultDocument[]
-
-/* ------------------------------------------------------------------------- */
-
-/**
  * Search result
  */
-export interface SearchResult {
-  items: SearchResultItem[]            /* Search result items */
-  suggestions?: string[]               /* Search suggestions */
-}
+export type SearchResult = Array<SearchDocument & SearchMetadata>
 
 /* ----------------------------------------------------------------------------
  * Functions
@@ -114,7 +113,7 @@ export interface SearchResult {
  * @param a - 1st list of strings
  * @param b - 2nd list of strings
  *
- * @return Difference
+ * @returns Difference
  */
 function difference(a: string[], b: string[]): string[] {
   const [x, y] = [new Set(a), new Set(b)]
@@ -129,9 +128,6 @@ function difference(a: string[], b: string[]): string[] {
 
 /**
  * Search index
- *
- * Note that `lunr` is injected via Webpack, as it will otherwise also be
- * bundled in the application bundle.
  */
 export class Search {
 
@@ -140,8 +136,8 @@ export class Search {
    *
    * A mapping of URLs (including hash fragments) to the actual articles and
    * sections of the documentation. The search document mapping must be created
-   * regardless of whether the index was prebuilt or not, as `lunr` itself will
-   * only store the actual index.
+   * regardless of whether the index was prebuilt or not, as Lunr.js itself
+   * only stores the actual index.
    */
   protected documents: SearchDocumentMap
 
@@ -151,24 +147,16 @@ export class Search {
   protected highlight: SearchHighlightFactoryFn
 
   /**
-   * The underlying `lunr` search index
+   * The underlying Lunr.js search index
    */
   protected index: lunr.Index
-
-  /**
-   * Search options
-   */
-  protected options: SearchOptions
 
   /**
    * Create the search integration
    *
    * @param data - Search index
    */
-  public constructor({ config, docs, index, options }: SearchIndex) {
-    this.options = options
-
-    /* Set up document map and highlighter factory */
+  public constructor({ config, docs, pipeline, index }: SearchIndex) {
     this.documents = setupSearchDocumentMap(docs)
     this.highlight = setupSearchHighlighter(config)
 
@@ -177,7 +165,7 @@ export class Search {
 
     /* If no index was given, create it */
     if (typeof index === "undefined") {
-      this.index = lunr(function() {
+      this.index = lunr(function () {
 
         /* Set up multi-language support */
         if (config.lang.length === 1 && config.lang[0] !== "en") {
@@ -189,7 +177,7 @@ export class Search {
         /* Compute functions to be removed from the pipeline */
         const fns = difference([
           "trimmer", "stopWordFilter", "stemmer"
-        ], options.pipeline)
+        ], pipeline!)
 
         /* Remove functions from the pipeline for registered languages */
         for (const lang of config.lang.map(language => (
@@ -231,9 +219,9 @@ export class Search {
    *
    * @param query - Query value
    *
-   * @return Search results
+   * @returns Search results
    */
-  public search(query: string): SearchResult {
+  public search(query: string): SearchResult[] {
     if (query) {
       try {
         const highlight = this.highlight(query)
@@ -248,7 +236,7 @@ export class Search {
         const groups = this.index.search(`${query}*`)
 
           /* Apply post-query boosts based on title and search query terms */
-          .reduce<SearchResultItem>((items, { ref, score, matchData }) => {
+          .reduce<SearchResult>((results, { ref, score, matchData }) => {
             const document = this.documents.get(ref)
             if (typeof document !== "undefined") {
               const { location, title, text, parent } = document
@@ -261,7 +249,7 @@ export class Search {
 
               /* Highlight title and text and apply post-query boosts */
               const boost = +!parent + +Object.values(terms).every(t => t)
-              items.push({
+              results.push({
                 location,
                 title: highlight(title),
                 text: highlight(text),
@@ -269,56 +257,34 @@ export class Search {
                 terms
               })
             }
-            return items
+            return results
           }, [])
 
           /* Sort search results again after applying boosts */
           .sort((a, b) => b.score - a.score)
 
           /* Group search results by page */
-          .reduce((items, result) => {
+          .reduce((results, result) => {
             const document = this.documents.get(result.location)
             if (typeof document !== "undefined") {
               const ref = "parent" in document
                 ? document.parent!.location
                 : document.location
-              items.set(ref, [...items.get(ref) || [], result])
+              results.set(ref, [...results.get(ref) || [], result])
             }
-            return items
-          }, new Map<string, SearchResultItem>())
+            return results
+          }, new Map<string, SearchResult>())
 
-        /* Generate search suggestions, if desired */
-        let suggestions: string[] | undefined
-        if (this.options.suggestions) {
-          const titles = this.index.query(builder => {
-            for (const clause of clauses)
-              builder.term(clause.term, {
-                fields: ["title"],
-                presence: lunr.Query.presence.REQUIRED,
-                wildcard: lunr.Query.wildcard.TRAILING
-              })
-          })
-
-          /* Retrieve suggestions for best match */
-          suggestions = titles.length
-            ? Object.keys(titles[0].matchData.metadata)
-            : []
-        }
-
-        /* Return items and suggestions */
-        return {
-          items: [...groups.values()],
-          ...typeof suggestions !== "undefined" && { suggestions }
-        }
+        /* Expand grouped search results */
+        return [...groups.values()]
 
       /* Log errors to console (for now) */
       } catch {
-        // tslint:disable-next-line no-console
         console.warn(`Invalid query: ${query} – see https://bit.ly/2s3ChXG`)
       }
     }
 
     /* Return nothing in case of error or empty query */
-    return { items: [] }
+    return []
   }
 }
