@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2020 Martin Donath <martin.donath@squidfunk.com>
+ * Copyright (c) 2016-2021 Martin Donath <martin.donath@squidfunk.com>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to
@@ -48,7 +48,7 @@ export type SearchQueryTerms = Record<string, boolean>
  *
  * @param value - Query value
  *
- * @return Search query clauses
+ * @returns Search query clauses
  */
 export function parseSearchQuery(
   value: string
@@ -67,7 +67,7 @@ export function parseSearchQuery(
  * @param query - Search query clauses
  * @param terms - Search terms
  *
- * @return Search query terms
+ * @returns Search query terms
  */
 export function getSearchQueryTerms(
   query: SearchQueryClause[], terms: string[]
