@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2020 Martin Donath <martin.donath@squidfunk.com>
+ * Copyright (c) 2016-2021 Martin Donath <martin.donath@squidfunk.com>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to
@@ -20,9 +20,6 @@
  * IN THE SOFTWARE.
  */
 
-export * from "./code"
-export * from "./details"
-export * from "./script"
+export * from "./indeterminate"
 export * from "./scrollfix"
-export * from "./source"
-export * from "./table"
+export * from "./scrolllock"
