@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2020 Martin Donath <martin.donath@squidfunk.com>
+ * Copyright (c) 2016-2021 Martin Donath <martin.donath@squidfunk.com>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to
@@ -20,32 +20,25 @@
  * IN THE SOFTWARE.
  */
 
-import { SourceFacts } from "patches/source"
-import { h, round } from "utilities"
+import { SourceFacts } from "~/components"
+import { h } from "~/utilities"
 
 /* ----------------------------------------------------------------------------
  * Functions
  * ------------------------------------------------------------------------- */
 
 /**
- * Render source facts
+ * Render repository facts
  *
- * @param facts - Source facts
+ * @param facts - Repository facts
  *
- * @return Element
+ * @returns Element
  */
-export function renderSource(
-  facts: SourceFacts
-) {
+export function renderSourceFacts(facts: SourceFacts): HTMLElement {
   return (
     <ul class="md-source__facts">
-      {Object.keys(facts).map(key => (
-        <li class={`md-source__fact md-source__fact--${key}`}>
-          {typeof facts[key as keyof SourceFacts] === "number"
-            ? round(facts[key as keyof SourceFacts])
-            : facts[key as keyof SourceFacts]
-          }
-        </li>
+      {facts.map(fact => (
+        <li class="md-source__fact">{fact}</li>
       ))}
     </ul>
   )
