@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2020 Martin Donath <martin.donath@squidfunk.com>
+ * Copyright (c) 2016-2021 Martin Donath <martin.donath@squidfunk.com>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to
@@ -20,16 +20,20 @@
  * IN THE SOFTWARE.
  */
 
-import { h } from "utilities"
+import { configuration } from "~/_"
+import { h } from "~/utilities"
 
 /* ----------------------------------------------------------------------------
  * Types
  * ------------------------------------------------------------------------- */
 
-interface Version {
-  version: string
-  title: string,
-  aliases: string[]
+/**
+ * Version
+ */
+export interface Version {
+  version: string                      /* Version identifier */
+  title: string                        /* Version title */
+  aliases: string[]                    /* Version aliases */
 }
 
 /* ----------------------------------------------------------------------------
@@ -37,27 +41,35 @@ interface Version {
  * ------------------------------------------------------------------------- */
 
 /**
- * Render a version dropdown
+ * Render a version selector
  *
  * @param versions - Versions
  *
- * @return Element
+ * @returns Element
  */
-export function renderVersion(base: string, versions: Version[]) {
-  const [, current] = base.match(/([^\/]+)\/?$/)!
-  const currentVersion =
+export function renderVersionSelector(versions: Version[]): HTMLElement {
+  const config = configuration()
+
+  /* Determine active version */
+  const [, current] = config.base.match(/([^/]+)\/?$/)!
+  const active =
     versions.find(({ version, aliases }) => (
       version === current || aliases.includes(current)
     )) || versions[0]
+
+  /* Render version selector */
   return (
     <div class="md-version">
       <span class="md-version__current">
-        {currentVersion.version}
+        {active.version}
       </span>
       <ul class="md-version__list">
         {versions.map(version => (
           <li class="md-version__item">
-            <a class="md-version__link" href={`${base}/../${version.version}/`}>
+            <a
+              class="md-version__link"
+              href={`${new URL(version.version, config.base)}`}
+            >
               {version.title}
             </a>
           </li>
