@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2020 Martin Donath <martin.donath@squidfunk.com>
+ * Copyright (c) 2016-2021 Martin Donath <martin.donath@squidfunk.com>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to
@@ -20,7 +20,5 @@
  * IN THE SOFTWARE.
  */
 
-export * from "./config"
-export * from "./jsx"
-export * from "./rxjs"
+export * from "./h"
 export * from "./string"
